requirements.txt
akismet==0.2.0
amqp==1.4.9
anyjson==0.3.3
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
billiard==3.3.0.22
bleach==1.4
celery==3.1.20
celerymon==1.0.3
certifi==2015.11.20.1
chardet==2.2.1
Django==1.8.19
django-allauth==0.19.1
django-appconf==1.0
django-braces==1.4.0
django-celery==3.1.10
django-celery-email==1.0.4
django-classy-tags==0.7.1
django-compressor==2.2
django-cookie-law==1.0.6
django-crispy-forms==1.4.0
django-debug-toolbar==1.4
django-filer==1.1.1
django-front==0.5.1
django-grappelli==2.7.3
django-haystack==2.4.1
django-inplaceedit==1.4.1
django-inplaceedit-extra-fields==0.6.1
django-mptt==0.8.0
django-polymorphic==0.8.1
django-recaptcha==1.0.4
django-reversion==1.8.1
docutils==0.11
easy-thumbnails==2.3
elasticsearch==1.9.0
feedparser==5.1.3
html5lib==0.999
Jinja2==2.7.3
kombu==3.0.33
langdetect==1.0.5
Markdown==2.4.1
MarkupSafe==0.23
mechanize==0.2.5
MySQL-python==1.2.5
oauthlib==1.0.3
ordereddict==1.1
Pillow==3.1.1
pyelasticsearch==1.4
Pygments==1.6
pyparsing==2.0.2
python-openid==2.2.5
pytz==2015.7
pyzmail==1.0.3
requests==2.3.0
requests-oauthlib==0.6.0
simplejson==3.8.1
singledispatch==3.4.0.3
six==1.10.0
Sphinx==1.2.2
sqlparse==0.1.18
tornado==4.3
Unidecode==0.4.19
urllib3==1.14
waitress==0.8.10
wheel==0.24.0
whitenoise==4.1.2
Whoosh==2.7.0
rdflib==4.2.2
rdflib-jsonld==0.4.0
dj-static==0.0.6
gunicorn==19.9.0
